add repository

+++ app/src/main/java/lab/justonebyte/moneysubuu/data/CategoryEntity.kt
@@ -7,6 +7,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "category_table")
 class CategoryEntity(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "id") val id:Double,
+    @ColumnInfo(name = "id") val id:Int,
     @ColumnInfo(name = "name") val name: String
 )
